+++ crates/ruma-client-api/src/knock/knock_room.rs
@@ -8,7 +8,7 @@ pub mod v3 {
     //! [spec]: https://spec.matrix.org/v1.19/client-server-api/#post_matrixclientv3knockroomidoralias
 
     use ruma_common::{
-        OwnedServerName, RoomId, RoomOrAliasId,
+        RoomId, RoomOrAliasId, ServerName,
         api::{auth_scheme::AccessToken, error::Error, response},
         metadata,
     };
@@ -42,7 +42,7 @@ pub mod v3 {
         ///
         /// When deserializing, the value is read from `via` if it's not missing or
         /// empty and `server_name` otherwise.
-        pub via: Vec<OwnedServerName>,
+        pub via: Vec<ServerName>,
     }
 
     /// Data in the request's query string.
@@ -51,13 +51,13 @@ pub mod v3 {
     struct RequestQuery {
         /// The servers to attempt to knock on the room through.
         #[serde(default, skip_serializing_if = "<[_]>::is_empty")]
-        via: Vec<OwnedServerName>,
+        via: Vec<ServerName>,
 
         /// The servers to attempt to knock on the room through.
         ///
         /// Deprecated in Matrix >1.11 in favour of `via`.
         #[serde(default, skip_serializing_if = "<[_]>::is_empty")]
-        server_name: Vec<OwnedServerName>,
+        server_name: Vec<ServerName>,
     }
 
     /// Data in the request's body.
@@ -169,7 +169,7 @@ pub mod v3 {
                 MatrixVersion, OutgoingRequestExt as _, SupportedVersions,
                 auth_scheme::SendAccessToken,
             },
-            owned_server_name, room_id,
+            room_id, server_name,
         };
 
         use super::Request;
@@ -177,7 +177,7 @@ pub mod v3 {
         #[test]
         fn serialize_request_via_and_server_name() {
             let mut req = Request::new(room_id!("!foo:b.ar").into());
-            req.via = vec![owned_server_name!("f.oo")];
+            req.via = vec![server_name!("f.oo")];
             let supported = SupportedVersions {
                 versions: [MatrixVersion::V1_1].into(),
                 features: Default::default(),
@@ -196,7 +196,7 @@ pub mod v3 {
         #[test]
         fn serialize_request_only_via() {
             let mut req = Request::new(room_id!("!foo:b.ar").into());
-            req.via = vec![owned_server_name!("f.oo")];
+            req.via = vec![server_name!("f.oo")];
             let supported = SupportedVersions {
                 versions: [MatrixVersion::V1_12].into(),
                 features: Default::default(),
@@ -215,7 +215,7 @@ pub mod v3 {
 
     #[cfg(all(test, feature = "server"))]
     mod tests_server {
-        use ruma_common::{api::IncomingRequestExt as _, owned_server_name};
+        use ruma_common::{api::IncomingRequestExt as _, server_name};
 
         use super::Request;
 
@@ -246,7 +246,7 @@ pub mod v3 {
 
             assert_eq!(req.room_id_or_alias, "!foo:b.ar");
             assert_eq!(req.reason, Some("Let me in already!".to_owned()));
-            assert_eq!(req.via, vec![owned_server_name!("f.oo")]);
+            assert_eq!(req.via, vec![server_name!("f.oo")]);
         }
 
         #[test]
@@ -263,7 +263,7 @@ pub mod v3 {
 
             assert_eq!(req.room_id_or_alias, "!foo:b.ar");
             assert_eq!(req.reason, Some("Let me in already!".to_owned()));
-            assert_eq!(req.via, vec![owned_server_name!("f.oo")]);
+            assert_eq!(req.via, vec![server_name!("f.oo")]);
         }
 
         #[test]
@@ -280,7 +280,7 @@ pub mod v3 {
 
             assert_eq!(req.room_id_or_alias, "!foo:b.ar");
             assert_eq!(req.reason, Some("Let me in already!".to_owned()));
-            assert_eq!(req.via, vec![owned_server_name!("f.oo")]);
+            assert_eq!(req.via, vec![server_name!("f.oo")]);
         }
     }
 }