+++ crates/ruma-common/tests/it/identifiers/ui/01-valid-id-macros.rs
@@ -8,8 +8,8 @@ fn main() {
     _ = ruma_common::room_id!("!1234567890:matrix.org");
     _ = ruma_common::room_version_id!("1");
     _ = ruma_common::room_version_id!("1-custom");
-    _ = ruma_common::server_signing_key_version!("Abc_1");
     _ = ruma_common::server_name!("myserver.fish");
+    _ = ruma_common::server_signing_key_version!("Abc_1");
     _ = ruma_common::user_id!("@user:ruma.io");
 
     _ = ruma_common::owned_user_id!("@user:ruma.io");
@@ -23,5 +23,6 @@ fn main() {
         _ = ruma_common::mxc_uri_ref!("mxc://myserver.fish/sdfdsfsdfsdfgsdfsd");
         _ = ruma_common::room_alias_id_ref!("#alias:server.tld");
         _ = ruma_common::room_id_ref!("!1234567890:matrix.org");
+        _ = ruma_common::server_name_ref!("myserver.fish");
     }
 }