+++ crates/ruma-events/src/space/parent.rs
@@ -2,7 +2,7 @@
 //!
 //! [`m.space.parent`]: https://spec.matrix.org/v1.19/client-server-api/#mspaceparent
 
-use ruma_common::{OwnedServerName, RoomId};
+use ruma_common::{RoomId, ServerName};
 use ruma_macros::EventContent;
 use serde::{Deserialize, Serialize};
 
@@ -18,7 +18,7 @@ use serde::{Deserialize, Serialize};
 #[ruma_event(type = "m.space.parent", kind = State, state_key_type = RoomId)]
 pub struct SpaceParentEventContent {
     /// List of candidate servers that can be used to join the room.
-    pub via: Vec<OwnedServerName>,
+    pub via: Vec<ServerName>,
 
     /// Determines whether this is the main parent for the space.
     ///
@@ -35,7 +35,7 @@ pub struct SpaceParentEventContent {
 
 impl SpaceParentEventContent {
     /// Creates a new `SpaceParentEventContent` with the given routing servers.
-    pub fn new(via: Vec<OwnedServerName>) -> Self {
+    pub fn new(via: Vec<ServerName>) -> Self {
         Self { via, canonical: false }
     }
 }
@@ -55,17 +55,15 @@ impl PossiblyRedactedSpaceParentEventContent {
 
 #[cfg(test)]
 mod tests {
-    use ruma_common::{canonical_json::assert_to_canonical_json_eq, owned_server_name};
+    use ruma_common::{canonical_json::assert_to_canonical_json_eq, server_name};
     use serde_json::json;
 
     use super::SpaceParentEventContent;
 
     #[test]
     fn space_parent_serialization() {
-        let content = SpaceParentEventContent {
-            via: vec![owned_server_name!("example.com")],
-            canonical: true,
-        };
+        let content =
+            SpaceParentEventContent { via: vec![server_name!("example.com")], canonical: true };
 
         assert_to_canonical_json_eq!(
             content,